Corrección de errores menores

+++ src/main/java/com/example/gestor_recetas/service/RecetaService.java
@@ -17,46 +17,41 @@ public class RecetaService {
     @Autowired // Instancia la clase, "inyección de depencendias"
     private RecetaRepository recetaRepository;
 
-    // Método como .save son del framework
-
     // POST
-    public Receta guadarReceta(Receta receta) {
+    public Receta guardarReceta(Receta receta) {
         return recetaRepository.save(receta);
     }
 
-    // GET ALL
-    // Rember list are declared List<varType> varName,
-    // and functions (have return) need a variable type when declaring
-
     // OBTENER TODOS
-    // Recuerda que las listas se declaran como List<tipoDeVariable> nombreVariable,
-    // y que las funciones (que tienen un valor de retorno) necesitan especificar un
-    // tipo de dato al declararse
     public List<Receta> obtenerTodos() {
         return recetaRepository.findAll();
     }
 
     // GET BY ID
-    // Se usa Integer ya que en la BD Esta con este termino y no LONG
+    // Se usa Integer ya que en la BD está como INT y no LONG
     public Optional<Receta> obtenerPorId(Integer id) {
         return recetaRepository.findById(id);
     }
 
     // PUT
-    public Receta actuReceta(Integer id, Receta detalles) {
+    public Receta actualizarReceta(Integer id, Receta detalles) {
         return recetaRepository.findById(id).map(recetaExistente -> {
             recetaExistente.setNombre(detalles.getNombre());
+            recetaExistente.setIngredientes(detalles.getIngredientes());
+            recetaExistente.setInstrucciones(detalles.getInstrucciones());
+            recetaExistente.setEnlace(detalles.getEnlace());
+            recetaExistente.setFuente(detalles.getFuente());
+            // recetaExistente.setNER(detalles.getNER());
+            return recetaRepository.save(recetaExistente);
+
         }).orElseThrow(() -> new ResourceNotFoundException("Receta no encontrada con ID: " + id));
-        // ^ Instanciar la clase creada en la carpeta excetion
     }
+
     // DELETE
-    public void eliminarReceta(Integer id){
-     // Se puede manejar la excepción con throw o try-catch
-     // Excepciones del runtime no exige de un try-catch para manejar la excepción
-        throw new ResourceNotFoundException("Receta no encontrada con ID: " + id);
+    public void eliminarReceta(Integer id) {
+        if (!recetaRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Receta no encontrada con ID: " + id);
+        }
+        recetaRepository.deleteById(id);
     }
-     // Notice it is byId, not All
-     recetaRepository.deleteById(id);
-
-
 }
